Add projects section to portfolio site
This commit will add the projects section to the portfolio website. The CSS will not be 100% responsive as of this commit. There will be one project card added so far. In the future there will still need to be styling changes as far as the text and animations go for this section. Possibly will need to add smooth scrolling feature for page transitions.